Store attribute filename in local backend's pdata
Previously, the filenames of the attributes were recomposed from
their name (e.g. the 'raw' attribute of the input channel 'magn_x'
would be give the filename 'in_magn_x_raw'); this did not work
for shared attributes.

With this commit, the filename of an attribute is saved in the
local backend's structure, so that there's no need to recompose it;
this also permits to read attributes shared among multiples channels.

This commit also introduces backend-provided functions to read/write
the channel attributes, as reading a device's attribute is a
different process than reading a channel's attribute.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>