Make iio_context_destroy free the devices
It was previously done in the backend-provided function; but as
free_device is now backend-agnostic, it's better to do it here.

Starting from now, the backend-provided 'shutdown' function
should only free the optional backend data of the context.

Additionally, the backend-provided 'shutdown' function can
now be omitted, if the backend has nothing specific to free.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>