Added a XML backend
The XML backend is meant as a debug tool. From the XML file whose
filename is passed as argument to the API function
iio_create_xml_context(), a full context is created.

This XML corresponds to a snapshot of a context, captured by
the Python tool 'iioxml.py' in the python sub-directory. The XML
backend restores the context from this snapshot.

While the general structure (name/number of devices, channels,
attributes) is restored, the backend of the context restored is
now the XML backend and not the local backend, and thus the
backend functions are different; Concretely, this means that
reading values of attributes will read their saved value from
the XML file and no more from sysfs; it also means that setting
new values to attributes is not possible.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>