Return -ENOSYS if the read/write backend functions don't exist
Previously, the API functions to read/write device/channel
attributes would always call the backend functions, which would
result in a segmentation fault if one of those was not implemented.

By first checking that the backend functions exist, we allow
backends not to support a given operation. For instance, the XML
backend does not handle writing attributes.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>