local backend: Fix the get_filename function
The get_filename function is meant to map a filename from a
channel's attribute. But before this commit, only the name of the
attribute was checked, not whether or not the channel in the map
is the one we are using.

This was not detected sooner because strangely enough, it was
working fine when used with the adis16488 board.

This fixes a bug that appeared in commit 3006910b.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>