Replace iio_channel_get_type by iio_channel_is_output
The API function iio_channel_get_type used to return an enum
with three possible values: input, output, unknown. However,
the local backend would create a iio_channel object only if
the type was known (either input or output), so the 'unknown'
type was actually never used.

The new API function iio_channel_is_output returns a boolean,
which indicates whether or not the channel is an output channel.
In case it is not, you can safely deduce that the channel is an
input channel.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>