Remove un-needed 'const' keywords
The name, id and attributes of channels and devices were stored
in their respective structure as const char *. However, all of
those fields are pointers to allocated data, which is not const.
This resulted in a high number of casts to the non-const type char*.

Having const fields in a structure can be justified when this
structure is visible on the public API, to inform that the data
pointed should not be modified. Here, the iio_device and iio_channel
structures have a limited visibility; having const fields is not
useful.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>