Local backend: Move the filename map out of context pdata
Having the filename map in channel pdata makes it much less
painful to search the filename of a channel's attribute. Previously,
the get_filename() function would search the unique big map stored
in context pdata; now it only search the map in channel data, which
contains only the filenames of the channel's own attributes. This
is faster, cleaner, better.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>