IIOD: Fixed two races
The first race was happening when a call to read_buffer was made
just a tiny bit after the read thread entered its shutdown phase;
in this case, iio_device_open() was sometimes called before
iio_device_close(), which returned an error.

The second race was happening when the read thread was signaling
back the waiting threads; those would instantly free the ThdEntry
structure, which was still needed in the read thread (to obtain the
address of the next element).

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>