IIOD: Added commands OPEN and CLOSE
Previously, the daemon would open and close the IIO devices on
demand (open when READBUF was used, close after a timeout).

However, we want to be able to select a mask of channels when
opening the device, so that we receive only the data of the channels
we're interested in.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>