IIOD: Register a client thread as soon as OPEN is used
Previously, a client thread was registered as a client for a device
at the beginning of the read_buffer function (called with the READBUF
command from the interpreter) and de-registered as soon as the read
was complete. This caused several problems; for instance, it was not
possible for the reading thread to know the best channel mask to use,
as it had no pointers to the structures of its clients.

Now, the client thread is registered when the OPEN command is used,
and it is de-registered only when CLOSE is used. This permits to
dynamically adjust the number of opened channels, according to the
channel mask of each client, when clients use those commands.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>