Change prototype of function iio_device_read_raw
Also change its backend versions.

The iio_device_read_raw will now return a channel mask corresponding
to the buffer that has been read.

This change is rather important in the network backend, as it drops
the client-side demuxing; I decided that this should not appear in the
"read" backend function but at a higher level.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>