Change prototype of API function iio_device_create_buffer
It was previously accepting a boolean parameter to inform of the type
of the buffer: input or output.

This was causing more trouble than solving issues. As internally the
IIO device is opened as soon as the iio_buffer is created, with the
local backend it was not possible to create an input buffer and an
output buffer at the same time for the same device.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>