Added API function iio_context_get_version
Also added constants in the public header iio.h to inform about the
version of the API programs are compiled with.

Then, it is easy to verify that the version of libiio which was used at
compilation time is the same than the one used at runtime.

Note that iio_context_get_version does not exactly return the version of
the library, but the version of the backend. For most backends, this
corresponds to the version of the library, but the network backend for
instance will instead return the version of the library used by the IIO
daemon.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>