Added three API functions
The three functions are:
iio_device_attr_write_raw
iio_device_debug_attr_write_raw
iio_channel_attr_write_raw

They can be used to write a buffer of data instead of a NULL-terminated
string to an attribute.

The functions iio_device_attr_write, iio_device_debug_attr_write and
iio_channel_attr_write now call these functions with a buffer size
corresponding to the length of the string (\0 included).

This change modifies the commands sent to the IIO daemon, but is needed
for the IIO daemon to implement the new commands (a sort of
chicken-and-egg situation), so the network backend is temporarily
broken.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>