Change prototype of backend function get_buffer
It was previously taking a channel mask as argument, which ended up
unused. As the local backend is the only backend which will ever
implement the get_buffer function, we can assume that the channel mask
will not change.

Instead, the prototype now takes a size_t parameter which corresponds to
the number of bytes present in the buffer that should be pushed to the
hardware. When using the high-speed mmap interface on an input device,
it is safe to set that value to zero; it is only useful with output
devices.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>