IIOD: Make VERSION return the version of libiio it was compiled with
Previously it would return the version of libiio it is currently using
at runtime, which may be different and newer.

If for instance a new version of the library introduce the foo()
function, returning the version of the library would trick the client
application into assuming that the IIO daemon supports the associated
"FOO" command, which is not true if the daemon was built from older
sources.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>