Local backend: Add a way to read all of the device/channel attributes
By passing NULL as the "attr" argument to iio_device_attr_read or
iio_channel_attr_read, it is now possible to read all of the
attributes of a device or channel.

The buffer is filled with one block of data per attribute of the
device or channel, by the order they appear in the iio_device/iio_channel
structure.

The first four bytes of one block correspond to a 32-bit signed value in
network order. If negative, it corresponds to the errno code that were
returned when reading the attribute; if positive, it corresponds to the
length of the data read. In that case, the rest of the block contains
the data.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>