IIOD: Fix a race occuring when the RW thread fails to create the buffer
Newly connected clients will wait from the open_dev function for a
signal from the RW thread, while the regular clients will wait from the
rw_buffer function. The difference is that in the former case, the
ThdEntry structure is freed as soon as the signal is received, so we
must consider the client's ThdEntry structure as not usable anymore as
soon as we wake up the client.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>