Local backend: Enable buffer before writing samples in non-cyclic mode
With the low-speed interface, on devices that require cyclic buffers,
the buffer has to be enabled after the samples are written.

However, on devices that do not use the cyclic mode but the streaming
mode, the buffer has to be enabled prior to writing the samples. This
commit makes sure that this requirement is respected.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>