Added API function iio_create_default_context
This function will create a network context if the IIOD_REMOTE
environment variable is set to the hostname where the IIOD server
runs. If set to an empty string, the server will be discovered using
ZeroConf. If the environment variable is not set, a local context
will be created instead.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>