iio_readdev: Don't access IIO API in signal handler
Accessing the IIO API from a signal handler is not safe and e.g. when using
the network backend can result in deadlocks or communication failures.  So
instead of cleaning up from within the signal handler add a variable that
says whether the application should be running or not and set it to false
from the signal handler. The main loop will then quit once it sees it being
false.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>