Network backend: Configure sockets with TCP_NODELAY
This ensures that the TCP segments are sent as soon as possible, even if
there is only a small amount of data.

As the communication between the network backend of libiio and IIOD is
mainly composed of tiny messages (< 10 bytes), this change speeds up
considerably the speed of communication between the two.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>