tests: Add iio_adi_xflow_check tool
The iio_adi_xflow_check tool calls either, depending on whether the
specified device is a transmit or receive device, iio_buffer_push() or
iio_buffer_refill() in a tight loop. Simultaneously it launches a monitor
thread that reads the status register from the ADI AXI ADC/DAC core and
checks the status register for underflows or overflows.

This can for example be used to debug and analyze memory bandwidth issues.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>