Revert "cmake: libxml2 detection: try first the CMake module from libxml2"
The CMake module provided by libxml2 does not work correctly, as it
sets up an incorrect list of include directories.

A patch has recently been sent upstream to address the issue:
https://mail.gnome.org/archives/xml/2014-December/msg00015.html

In the meantime, it is safer to revert this commit for now, and restore
it when the fix is available in an official release of libxml2.

This reverts commit 6014cd588ec3c9d8daa513ab0c943957021d0df8.