local: Remove modifier detection from set_channel_name()
set_channel_name() detects a common prefix in the names of the attributes of
a channel and assigns this as the channel name. If it detects a common
prefix it will check if the prefix starts with a modifier name and use it to
setup the modifier of the channel. But when a attribute name is initialized
we already skip past the modifier, which means that a channel name never
starts with a modifier and the check in set_channel_name() is always false.
So remove it.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>