buffer: Only de-queue buffer in iio_device_create_buffer() in tx mode
In high-speed mode when creating a buffer in transmit mode we need to
de-queue a block and initialize buf->buffer so a application can fill the
buffer with data prior to calling iio_buffer_push(). In receive mode on the
other hand the application is expected to call iio_buffer_refill() before
accessing the data. iio_buffer_refill() will de-queue another block. This
means that the block that was de-queued in iio_device_create_buffer() will
be silently discarded and the data in it is lost.

In order to avoid this make sure to only de-queue a block in
iio_device_create_buffer() when running in transmit mode.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>