Added properties support for context factories
Some backends must be parametrized before their context is created. It's the
case for the network backend which needs an hostname.
So this patch proposes a generic yet simple way to do so with properties. They
are key/value string pairs, attached to a factory, which can then retrieve their
value and adapt the context creation process accordingly.
The iio_property is exposed in the public API, because it will be used by the
client program, for parametrizing the construction of the context it is
interested in.

Change-Id: I761462450e0abb4f3ccd6e880a37291b159c293d