Added implementation of context factory (un)registration
This mechanism allows to register a function for creating a context, in
a generic way. It is intended to be used for backend implementations.
This patch only implement the factory registration / unregistration functions,
which isn't useful by itself until some backend implementation actually uses
it.
A static array of factories, with a limited size, keeps track of all the
factories registered so far.

Change-Id: I0533490532e40f92ceef113314f4f4098d79aad9