Export the API functions needed by plugins
Plugin implementors willing to add for example, the support for a new backend,
must access some internal APIs. This patch prevents some functions from being
hidden. The functions concerned are those making possible to compile the local
backend as an independent plugin.
Note: they are not exposed as a public API in the iio.h file because they should
only be exposed to plugin implementors who will have to include the
iio-private.h header and certainly also debug.h.

Change-Id: Ib8d202cebe0ec8302d4fa5f7bd77ba079f2cf516