Local backend: Create context with it's factory
This patch implements a way to actually retrieve a factory, knowing it's name
and uses it for the local backend's creation.
Registration for the local backend's context factory is done in it's constructor
initialization function, thus saving the need of an extra function call at
startup and above all, making the context factories registrations independant of
the actual list of backends implemented. This opens the way for plugins
implementation.
The local.c file is compiled iif LOCAL_BACKEND is defined, the construction of
it's context will be successful if it's factory has already been registered and
will return NULL if not.

Change-Id: I134a0d17da7742475587b1597c69a1085e175841