Change prototype of backend function "open"
Previously, a channel mask was passed to the "open" function. The point
of this was that the open() function could modify the mask, for instance
in the case where channels cannot be disabled.

This did not work, as the Buffer object uses the device's channel mask
to know which samples should be made available to the application.

To handle those cases, the backends should instead modify the mask
passed to the read() and/or get_buffer() backend functions.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>