Local backend: Avoid warning from valgrind about uninitialized param
The parameter passed to ioctl() is a pointer to a structure that is
filled by the kernel, so it doesn't matter if the data structure is
initialized or not.

But valgrind doesn't know that, so we make sure that all fields of the
data structure are initialized before calling ioctl().

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>