Revert "CMake: Fix target name different from project name"
This is actually normal that the library is called 'lib<name>.a' under
Linux and just '<name>.dll' under Windows, and has nothing to see with
the project name.

However, by changing the target name to 'libiio' instead of just 'iio',
the CMake build system was creating the library as 'liblibiio.a'.

This reverts commit 8c46d527eba891e3cd437a73af807ac03a9c6cbf.