IIOD: Memorize if the input/output FDs are sockets or not
This optimizes the case where the input FD or the output FD is not a
socket. The code won't try to use socket-specific calls everytime
(send/recv), but only once, and if those calls fail with the ENOTSOCK
return code ("not a socket" error), then IIOD will use standard I/O
calls for the rest of the session.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>