Allow creating a USB context from iio_create_default_context()
By setting IIOD_REMOTE to a vid:pid couple, it is now possible to create
a USB context using iio_create_default_context().

As a result, all the applications that were using that function to
create the context will work out-of-the-box with the USB context (as
long as they don't rely on specific functionalities being present
and handle errors properly).

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>