IIOD: Fix bug introduced in 33613f2
The optimization introduced in commit 33613f2 ("IIOD: Improve parser
input function") came with a side effect. Since we were not reading
character-by-character anymore, there was no guarantee that the readfd()
function would only return the bytes composing the command, until the
trailing \n. In practice, that was always the case, except (in very
rare occasions) for with the WRITEBUF command, as the client might send
the data right after the command and before reading any error code. As
a result, the recv() call for the WRITEBUF command would read more
bytes than just the command, thus disturbing the communication and
crashing the IIO Daemon.

Instead of reverting to reading character-by-character, the new code
will first read one buffer of data without advancing the read offset,
search for the trailing \n, and advance the read offset by the correct
amount of bytes.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>