local: Detect channels without index
Add a second pass to detect_and_move_global_attrs() that detects channels
without an index and creates the channel.

We can't do this earlier since we can't distinguish from channels without
an index and shared attributes of channels with an index. So we only create
a channel without an index for attributes after the pass that detects and
assigns shared attributes.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>