network: Always handle EINTR
We need to always correctly handle EINTR and retry the socket operation
when we see it. Otherwise if we are in the middle of a recv() or send() and
it gets interrupted the communication flow between the server and the
client might get irreparably disrupted. The only way to recover in such a
scenario would be to destroy and re-create the context.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>