iio-monitor: Refresh monitor screen in one go
Avoid flicker in the monitor screen by performing only a single refresh per
update. Currently 3 refreshes are done, one by eraseCursesWindow(), one by
boxWindow() and finally the explicit refresh by wrefresh().
eraseCursesWindow() is replaced with werase() which is the same, except
that it does not refresh, and the boxWindow() is moved to the end after all
updates have been done and replaces the explicit wrefresh().

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>