Introduce URI based way for creating contexts
Currently applications that want to create a context other than the default
context need explicit support for each kind of backend type they want to
support and also need to know how each individual backend is addressed.
This does not scale well as new backends are introduced. Before the backend
can be used in a application support for it needs to be added to the
application. In addition this also leads to a lot of duplicated code since
applications are likely to use similar approaches to instantiating the
backends.

This patch introduces a new URI based way of creating the backend. Parsing
of the URI is done inside the libiio and applications merely have to
forward a user supplied string. This means whenever libiio is updated to
support new backends all applications that instantiate the backend by a URI
will automatically have support for it.

The URI is composed of a scheme part followed by an address part. Currently
the following URI schemes are supported:
	* 'local:', creates a local context
	* 'ip:[address]', creates a network context for the device at the
	  specified network address.
	* 'usb:[bus].[address]', creates a USB backend for the device at the
	   specified USB bus and address.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>