iio-private.h: Define the __api token here too
There is one private function () which is declared as __api.
The IIO daemon needs this function to be exported, but we don't want it
in the public iio.h file. Since the previous commit undefined the __api
at the end of the public iio.h header, we need to redeclare it here.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>