Revert "IIOD: Refill the buffer when no clients, if we had clients just before"
This optimization was useful when it was introduced, in May 2014.
It was based on the assumption that if the RW thread doesn't have
readers, but had readers in the previous iteration, then it's highly
likely that some clients will ask for data in this iteration as well,
while we are refilling the buffer.

However, this introduced a corner case that caused the IIOD to drop one
buffer of samples, if no clients asked for data during the refill.
This typically happens when using the high-speed buffer interface in the
local backend, and with a sample rate higher than what the network can
deliver, since then refilling the buffer hardly takes any time.
This went unnoticed, mostly because in that typical use case, a lot of
samples are dropped in the kernel anyway, as the DMA copies faster than
we can refill.

While we don't refill and handle clients in parallel anymore, this
revert shouldn't have implications on the performance. When using a
high-speed IIO device, as soon as one or more clients ask for data,
the refill will have high chances to return without sleeping as data
becomes available. With a low-speed IIO device, the added delay should
not be a problem due to the low data rate.

This reverts commit 8468985, and fixes issue #70.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>