iiod: Fix pthread condition handling
pthread conditions are supposed to be used in combination with a predicate
and a lock that protects the predicate, otherwise they are prone to race
conditions and/or deadlocks since the waiter might end the signal if it
starts waiting after the signal has been sent or spurious wake ups might
cause it to wake up to early.

Use the 'wait_for_open' and 'active' fields of the ThdEntry struct as the
predicates for the open and read/write operations. The DevEntry's
thdlist_lock is used as the lock to protect the predicates since we are
already pretty much always holding it when they are accessed and in cases
where we don't they are release right before accessing the predicates.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>