iiod: Remove last_cond
The race condition it tries to protect against does not exist. The
devlist_lock is being held when the thread is shutdown and the buffer
destroyed. This means a new client trying to open the device will block
until that operation has completed since it also takes the devlist_lock.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>