iiod: Zero alloc thread and device entry structs
Make sure that all fields that are not explicitly initialized are set to 0.
Otherwise undefined behaviour might occur. E.g. if active is set to non 0,
another client is already reading and the new client gets added to the
client list just at the right moment it will get send data without having
requested it.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>