iiod: Introduce per client list of open devices
Currently open devices are only tracked as part of the RW thread they are
attached to. If the thread aborts (e.g. due to an error) all references to
the open devices are lost and their data is never freed. To avoid this add
a list for each client which keeps track of the open devices.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>