iiod: open_dev_helper(): Alloc ThdEntry outside the locked section
The ThdEntry struct allocation does not need to be protected by a lock. So
move it outside of the locked section, this makes the flow a bit nicer.

The only downside is that now if a new client connects in cyclic mode and
another one is already active we have free the struct again without ever
using it. But since this should be the exception it should be OK.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>