iiod: Cleanup on shutdown
When a signal is sent to the iiod it will be handled by a random thread.
When running in interactive mode the signal handler will call exit() and
the iiod will stop immediately without doing any cleanup work.

When running in non-interactive mode the behaviour depends which thread the
signal is delivered to, when being delivered to one of the worker threads
it might ignore the signal or it might exit the thread. All other threads
will keep running either way. When a signal is received by the iiod the
main thread will abort and eventually exit while all other threads continue
running normally. This has multiple issues:
	* Active buffers wont be disabled
	* The main thread destroys the IIO context and might cause a
	  use-after-free in one of the worker threads which is still accessing
	  the context

In addition to that it is also possible for signals to be lost if the main
thread is not waiting when the signal is delivered.

So overall the behavior is unpredictable and in the best case will leak
resources and in the worst case crash the application.

To address this a new eventfd is introduced. It is used to signal that the
iiod has received a request to shutdown. All places where blocking IO is
performed are replaced with code that either waits for a shutdown event or
for the socket/file descriptor to become ready. If a shutdown event is
received the worker threads will cleanly shutdown and release all their
resources. In addition the main thread will wait until all worker threads
have finished running before destroying the IIO context and exiting the
application.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>