iiod: Add AIO support
In order to be able to properly shutdown iiod cleanly we need to make sure
that we can wakeup the RW thread when a shutdown request is received. When
using blocking IO this is only possible by using signals which is a bit
ugly and difficult to get implemented completely race condition free. So we
use non-blocking IO and poll() to be able to wait for both data and the
shutdown event.

Unfortunately it is not possible to use non-blocking IO on all file
descriptors (e.g. the USB gadgetfs endpoints). As an alternative to
non-blocking IO introduce support for AIO which allows the application to
submit IO requests and wait for the completion on a eventfd. If a shutdown
request is received while the IO operation is in progress the IO operation
is cancelled using the AIO interface.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>